+++ vitapharm-client/package.json
@@ -1,54 +1,102 @@
 {
 
   "proxy": "http://127.0.0.1:5000",
-  "name": "vitapharm",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@chakra-ui/icons": "^2.1.1",
-    "@chakra-ui/react": "^2.8.2",
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.5",
-    "@fortawesome/fontawesome-svg-core": "^6.5.2",
-    "@fortawesome/react-fontawesome": "^0.2.0",
-    "date-fns": "^3.6.0",
-    "date-fns-tz": "^3.1.3",
-    "framer-motion": "^11.1.9",
-    "js-cookie": "^3.0.5",
-    "react": "^18.2.0",
-    "react-burger-menu": "^3.0.9",
-    "react-cookie": "^7.1.4",
-    "react-dom": "^18.2.0",
-    "react-feather": "^2.0.10",
-    "react-hook-form": "^7.51.4",
-    "react-select": "^5.8.0",
-    "react-select-country-list": "^2.2.3",
-    "react-social-media-embed": "^2.5.13",
-    "react-text-transition": "^3.1.0",
-    "swiper": "^11.1.1",
-    "uuid": "^9.0.1"
-  },
-  "devDependencies": {
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
-    "@vitejs/plugin-react": "^4.2.1",
-    "autoprefixer": "^10.4.19",
-    "eslint": "^8.57.0",
-    "eslint-plugin-react": "^7.34.1",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.6",
-    "postcss": "^8.4.38",
-    "react-router-dom": "^6.22.3",
-    "tailwindcss": "^3.4.3",
-    "vite": "^5.2.0"
-  }
-}
 
+  "name": "vitapharm",
+
+  "private": true,
+
+  "version": "0.0.0",
+
+  "type": "module",
+
+  "scripts": {
+
+    "dev": "vite",
+
+    "build": "vite build",
+
+    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
+
+    "preview": "vite preview"
+
+  },
+
+  "dependencies": {
+
+    "@chakra-ui/icons": "^2.1.1",
+
+    "@chakra-ui/react": "^2.8.2",
+
+    "@emotion/react": "^11.11.4",
+
+    "@emotion/styled": "^11.11.5",
+
+    "@fortawesome/fontawesome-svg-core": "^6.5.2",
+
+    "@fortawesome/react-fontawesome": "^0.2.0",
+
+    "date-fns": "^3.6.0",
+
+    "date-fns-tz": "^3.1.3",
+
+    "framer-motion": "^11.1.9",
+
+    "js-cookie": "^3.0.5",
+
+    "react": "^18.2.0",
+
+    "react-burger-menu": "^3.0.9",
+
+    "react-cookie": "^7.1.4",
+
+    "react-dom": "^18.2.0",
+
+    "react-feather": "^2.0.10",
+
+    "react-hook-form": "^7.51.4",
+
+    "react-select": "^5.8.0",
+
+    "react-select-country-list": "^2.2.3",
+
+    "react-social-media-embed": "^2.5.13",
+
+    "react-text-transition": "^3.1.0",
+
+    "swiper": "^11.1.1",
+
+    "uuid": "^9.0.1"
+
+  },
+
+  "devDependencies": {
+
+    "@types/react": "^18.2.66",
+
+    "@types/react-dom": "^18.2.22",
+
+    "@vitejs/plugin-react": "^4.2.1",
+
+    "autoprefixer": "^10.4.19",
+
+    "eslint": "^8.57.0",
+
+    "eslint-plugin-react": "^7.34.1",
+
+    "eslint-plugin-react-hooks": "^4.6.0",
+
+    "eslint-plugin-react-refresh": "^0.4.6",
+
+    "postcss": "^8.4.38",
+
+    "react-router-dom": "^6.22.3",
+
+    "tailwindcss": "^3.4.3",
+
+    "vite": "^5.2.0"
+
+  }
+
+}
